changes produced by the following prompt:
Write a content script for a chrome extension that automatically determine and set the writing direction of the focused input element.
In addition, the user will be able to override the writing direction, meaning to set it to rtl if the right control and the right shift keys are pressed, and to ltr if the left control and left shift keys are pressed.
When a writing direction is set by the extension, it should should also align the text alignment to "start".
It should log to the console whenever a direction switch occurs.
It should change the writing direction only if the focused element is editable.
The automatic direction ability will work as follows: while the user is typing inside a focused input field, check what is the first word character (ignore non-word characters) and if it's a word character of an RTL language, set the text direction accordingly, or set it to LTR otherwise.
If the user had set the text direction explicitly using the shortcut, stop using the auto-direction ability.
If the user cleared the input field, re-enable the auto detection mode.
Use the input field itself to store the writing direction state, because a webpage can contain multiple input fields.
In general, avoid writing repetitive duplicated code as much as possible.
In general, avoid loops whenever possible.
In general, use /\p{L}/u unicode regular expressions, to match word characters.