changes produced by the following prompt:
Add a "Use Cases" section and describe that on Windows, users are able to change the writing direction to LTR\RTL using CTRL + LEFT\RIGHT SHIFT shortcuts. This is something that is missing from Chrome on Linux and that this extension enables it. Also, this extension automatically sets the writing direction in webpages that align the text to a fixed side - for example, In Google's website, the search field is fixed